requirements.txt
streamlit>=1.28.0
requests>=2.31.0
twelvelabs>=0.2.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.10
pytest>=7.4.0
pytest-cov>=4.1.0
